refactor(Clipboard copy related components): docs and adjustments
