+++ src/components/ClipboardIcon/README.md
@@ -1,0 +1,11 @@
+## ClipboardIcon
+
+This display **svg** copy-paste icon.<br/>
+Depending on **status** property additionally draws animated successful (✓ sing) or failed operation indicators.<br/>
+Might be adjusted by **className** and **size** props according to your needs.
+
+| Property  | Type     | Required | Default | Description                                                                         |
+| :-------- | :------- | :------: | :------ | :---------------------------------------------------------------------------------- |
+| status    | Enum     |          |         | Defines icon displayed: pending - default, <br/>success/error - additional drawings |
+| size      | `Number` |          | 24      | Icon size                                                                           |
+| className | `String` |          |         | CSS class name                                                                      |
