+++ src/components/CopyToClipboard/index.ts
@@ -1,1 +1,2 @@
 export * from './CopyToClipboard';
+export * from './types';
